requirements.txt
# Core dependencies
langchain==0.0.314
langchain-community==0.0.16
langchain-core==0.1.4
python-dotenv==1.0.0

# LLM APIs
openai==1.3.0
cohere-api==1.6.0
anthropic==0.5.0
google-generativeai==0.3.1

# Connectors
python-telegram-bot==20.6
requests==2.31.0
tweepy==4.14.0  # For general social media integration

# Vector databases & embeddings
chromadb==0.4.17
sentence-transformers==2.2.2

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2

# UI
gradio==3.50.2

# Testing and development
pytest==7.4.3
black==23.11.0

# Evaluation
ragas==0.0.20  # Para evaluación de respuestas 